+++ extensions/make4ht-ext-mathjaxnode.lua
@@ -7,8 +7,14 @@ function M.test(format)
   return true
 end
 
+function M.prepare_parameters(params)
+  params.tex4ht_sty_par = params.tex4ht_sty_par  .. ",mathml"
+  return params
+
+end
 function M.modify_build(make)
   local mathjax = filter { "mathjaxnode"}
+  -- this extension needs mathml enabled
   make:match("html?$",mathjax)
   return make
 end